Reference

User interface functions

ConstraintSolver.init — Function.

init()

Initialize the constraint model object

ConstraintSolver.add_var! — Function.

add_var!(com::CS.CoM, from::Int, to::Int; fix=nothing)

Adding a variable to the constraint model com. The variable is discrete and has the possible values from,..., to. If the variable should be fixed to something one can use the fix keyword i.e add_var!(com, 1, 9; fix=5)

ConstraintSolver.add_constraint! — Function.

add_constraint!(com::CS.CoM, constraint::Constraint)

Add a constraint to the model i.e add_constraint!(com, a != b)

ConstraintSolver.set_objective! — Function.

set_objective!(com::CS.CoM, sense::Symbol, objective::ObjectiveFunction)

Set the objective of the model. Sense can be :Min or :Max.

ConstraintSolver.solve! — Function.

solve!(com::CS.CoM; backtrack=true, max_bt_steps=typemax(Int64), backtrack_sorting=true, keep_logs=false)

Solve the constraint model if backtrack = true otherwise stop before calling backtracking. This way the search space can be inspected after the first pruning step. If max_bt_steps is set only that many backtracking steps are performed. If backtrack_sorting is set to false the same ordering is used as when used without objective this has only an effect when an objective is used. With keep_logs one is able to write the tree structure of the backtracking tree using ConstraintSolver.save_logs.

ConstraintSolver.value — Function.

value(v::CS.Variable)

Get the value of the variable if it is fixed. Otherwise one of the possible values is returned. Can be used if the status is :Solved as then all variables are fixed.

Constraints

These can be used as constraints for a model

ConstraintSolver.all_different — Method.

all_different(variables::Vector{Variable})

Create a BasicConstraint which will later be used by all_different(com, constraint).

Can be used i.e by add_constraint!(com, CS.all_different(variables)).

Base.:== — Method.

Base.:(==)(x::LinearVariables, y::Int)

Create a linear constraint with LinearVariables and an integer rhs y.

Can be used i.e by add_constraint!(com, x+y = 2).

Base.:== — Method.

Base.:(==)(x::LinearVariables, y::Variable)

Create a linear constraint with LinearVariables and a variable rhs y.

Can be used i.e by add_constraint!(com, x+y = z).

Base.:== — Method.

Base.:(==)(x::LinearVariables, y::LinearVariables)

Create a linear constraint with LinearVariables on the left and right hand side.

Can be used i.e by add_constraint!(com, x+y = a+b).

ConstraintSolver.equal — Method.

equal(variables::Vector{Variable})

Create a BasicConstraint which will later be used by equal(com, constraint)

Can be used i.e by add_constraint!(com, CS.equal([x,y,z]).

Base.:== — Method.

Base.:(==)(x::Variable, y::Variable)

Create a BasicConstraint which will later be used by equal(com, constraint)

Can be used i.e by add_constraint!(com, x == y).

Base.:! — Method.

Base.:!(bc::CS.BasicConstraint)

Change the BasicConstraint to describe the opposite of it. Only works with a equal basic constraint.

Can be used i.e by add_constraint!(com, x != y).

Objective functions

ConstraintSolver.vars_max — Method.

vars_max(vars::Vector{CS.Variable})

Create an objective which computes the maximum value a variable has if all are fixed otherwise depends on whether the objective sense is set to :Min or :Max.

Can be used i.e by set_objective!(com, :Min, CS.vars_max([x,y]).